Created rowGrabber.h. Pulled checker function out of sudoku.c into its own header file
